+++ comp2013/src/main/java/comp2013/IMovable.java
@@ -1,16 +1,9 @@
 package comp2013;
 
-import javafx.scene.image.Image;
-
 /**
- * 
- * @Project Snakee
- * @Description Fyrsta viðmótið
- * @Author Sigurður Sigurðardóttir
- * @version Ekki viss
- */ 
-
-public interface Movable
+ * Represents a movable snake object.
+ */
+public interface IMovable
 {
 	public int m_Direction = 0;